major refactoring to make the wrapper as smooth as possible
The function to read stuff doesn't look for a sentinal but instead
reads until no data has been received for a certain amount of time.
This means it works properly for everything zork can throw at us,
even if it doesn't end in a '>' prompt. Also the main loop was
restructured so that commands can be silently executed without the
user knowing (demonstrated as an auto save currently).